fix: Moved download_component to root
