--- cli/src/anipy_cli/comps/__init__.py
@@ -1,3 +1,0 @@
-from anipy_cli.comps.download_component import DownloadComponent
-
-__all__ = ["DownloadComponent"]
